Added ChaseState, Setup FSM

+++ C++ AI/RaylibStarterCPP/Player.cpp
@@ -47,8 +47,8 @@ void Player::DrawTextureWithPivot(const Texture& tex, Vector2 pos, Vector2 pivot
 {
 	Vector2 offset = pivot * -1;
 
-	float cosTheta = cos(rotation * DEG2RAD);
-	float sinTheta = sin(rotation * DEG2RAD);
+	float cosTheta = cos((double)rotation * DEG2RAD);
+	float sinTheta = sin((double)rotation * DEG2RAD);
 
 	Vector2 rotatedOffset;
